requirements.txt
websocket-client>=0.57.0
colorama>=0.4.4
requests>=2.25.1